+++ project/Build.scala
@@ -154,6 +154,15 @@ object BijectionBuild extends Build {
     libraryDependencies += jsonParser
   ).dependsOn(bijectionCore % "test->test;compile->compile")
 
+  lazy val bijectionAlgebird = Project(
+    id = "bijection-algebird",
+    base = file("bijection-algebird"),
+    settings = sharedSettings
+  ).settings(
+    name := "bijection-algebird",
+    libraryDependencies += "com.twitter" %% "algebird" % "0.1.6"
+  ).dependsOn(bijectionCore % "test->test;compile->compile")
+
   lazy val bijectionUtil = Project(
     id = "bijection-util",
     base = file("bijection-util"),